+++ browser/index.build.js
@@ -6,5 +6,5 @@
 // require('../third-party/bluebird.min.js');
 // require('../third-party/bignumber.min.js');
 
-import '../styles/febsui.scss';
+require('../styles/febsui.scss');
 require('./index');